% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/false_ring_proportion.R
\name{afrp}
\alias{afrp}
\title{adjusted false ring proportion}
\usage{
afrp(iadf)
}
\arguments{
\item{iadf}{A data frame with numeric columns representing individual series
and years as rownames where years with IADF are marked with 1,
those without with 0, years not covered by the series are set to NA.}
}
\value{
a data frame
}
\description{
Calculate the adjusted false ring proportion, as
  suggested by Osborn et. al. (1997), of a set of binary false ring assignments.
}
\references{
Osborn TJ, Briffa KR and Jones PD (1997) Adjusting variance for
  sample-size in tree-ring chronologies and other regional-mean timeseries.
  Dendrochronologia 15, 89-99.
}
\seealso{
\code{\link{frp}}
}
