% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/false_ring_proportion.R
\name{campelo_chapman_find_start}
\alias{campelo_chapman_find_start}
\title{campelo_chapman_find_start}
\usage{
campelo_chapman_find_start(
  campelo_freq_object,
  min.n = 15,
  max_a = 3,
  max_b = 1,
  max_c = 17
)
}
\arguments{
\item{campelo_freq_object}{a campelo frequency object,
output of \code{\link[iadf]{campelo_freq}}}

\item{min.n}{minimum number of samples within each group to be included in
model estimation}

\item{max_a}{maximum value of manipulate slider for parameter a}

\item{max_b}{maximum value of manipulate slider for parameter b}

\item{max_c}{maximum value of manipulate slider for parameter c}
}
\value{
a list which can be used as input argument 'start' in
\code{\link{campelo_chapman}}
}
\description{
Find good start values manually in case
\code{\link{campelo_chapman}}
  returns an error caused by insufficient default starting values.
}
