% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_iadf}
\alias{example_iadf}
\title{example_iadf}
\format{A data.frame with 135 years and 30 series.}
\usage{
example_iadf
}
\description{
An rwl object to be used in documented examples
}
\keyword{datasets}
