% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/false_ring_proportion.R
\name{frp}
\alias{frp}
\title{false ring proportion}
\usage{
frp(iadf)
}
\arguments{
\item{iadf}{A data frame with numeric columns representing individual series
and years as rownames where years with IADF are marked binary with 1,
those without with 0, years not covered by the series are set to NA.}
}
\value{
a data frame
}
\description{
Calculate the false ring proportion of a set of
  binary false ring assignments.
}
\seealso{
\code{\link{afrp}}
}
