% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/false_ring_proportion.R
\name{novak_negexp}
\alias{novak_negexp}
\title{novak_negexp}
\usage{
novak_negexp(novak_freq_object, min.n, start = NULL, make.plot = TRUE, ...)
}
\arguments{
\item{novak_freq_object}{A novak_freq_object as obtained from
\code{\link[iadf]{novak_freq}}}

\item{min.n}{minimum number of samples within each cambial age to be included in
model estimation}

\item{start}{set custom start values - default to \code{list(a = 0.01, b=1, c=1)}}

\item{make.plot}{logical}

\item{...}{additional plotting arguments}
}
\value{
a model object of class "nls"
}
\description{
This function fits a negative exponential function as an alternative
  to the Weibull function suggested by Novak et al. (2013) for calculation of
  age corrected IADF frequencies
}
\examples{
data('example_iadf')
model <- novak_negexp(novak_freq(example_iadf), 15)
novak_index(example_iadf, model)
}
\references{
Novak, Klemen and Sánchez, Miguel Angel Saz and Čufar, Katarina and Raventós,
  Josep and de Luis, Martin, te and intra-annual density fluctuations in in Spain,
  IAWA Journal, 34, 459-474 (2013), DOI:https://doi.org/10.1163/22941932-00000037
  \url{http://booksandjournals.brillonline.com/content/journals/10.1163/22941932-00000037}.
}
\seealso{
\code{\link{novak_freq}}, \code{\link{novak_weibull}},
  \code{\link{novak_index}}
}
