% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/false_ring_proportion.R
\name{novak_negexp_find_start}
\alias{novak_negexp_find_start}
\title{novak_negexp_find_start}
\usage{
novak_negexp_find_start(novak_freq_object, min.n = 15, max_a = 1,
  max_b = 3, max_c = 3)
}
\arguments{
\item{novak_freq_object}{A novak_freq_object as obtained from
\code{\link[iadfboku]{novak_freq}}}

\item{min.n}{minimum number of samples within each cambial age to be included in
model estimation}

\item{max_a}{maximum value of manipulate slider for parameter a}

\item{max_b}{maximum value of manipulate slider for parameter b}

\item{max_c}{maximum value of manipulate slider for parameter c}
}
\value{
a list which can be used as input argument 'start' in \code{\link{novak_weibull}}
}
\description{
Find good start values manually in case \code{\link{novak_negexp}}
  returns an error caused by insufficient default starting values
}
