% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{sort_by_index}
\alias{sort_by_index}
\title{sort_by_index}
\usage{
sort_by_index(x)
}
\arguments{
\item{x}{a numeric vector, representing an individual rwl series,
potentially containing NA values.}
}
\value{
a numeric vector with the same length as x.
}
\description{
internal function such as sortByIndex as in package dplR, shifts
  series to start with index 1, maintaining the same vector length by
  adding NA values to the end.
}
\examples{
x <- c(NA,NA,NA,1,2,3,4,5, NA, NA)
iadf:::sort_by_index(x)
#[1]  1  2  3  4  5 NA NA NA NA NA
}
